refactor: move file
This moves the navigation file outside of epic1 as it is to be used for all epics
